+++ Assets/Plugins/Substance 3D Extensions/Examples/Scripts/Runtime/Runtime/Controls/Logic/Float2InputControlLogic.cs
@@ -1,17 +1,14 @@
-using System.Collections;
-using System.Collections.Generic;
+using UnityEngine;
 using Adobe.Substance;
 using Adobe.Substance.Input;
-using Unity.VisualScripting.YamlDotNet.Core.Tokens;
-using UnityEngine;
 
 namespace SOS.SubstanceExtensions.Examples
 {
     public class Float2InputControlLogic : InputControlLogic<Vector2>
     {
-        public override void Initialize(SubstanceNativeGraph nativeGraph, SubstanceGraphSO substance, ISubstanceInput input)
+        public override void Initialize(SubstanceNativeGraph nativeGraph, ISubstanceInput input)
         {
-            base.Initialize(nativeGraph, substance, input);
+            base.Initialize(nativeGraph, input);
 
             SubstanceInputFloat2 valueInput = (SubstanceInputFloat2)input;