+++ Assets/Plugins/Substance 3D Extensions/Examples/Scripts/Runtime/Runtime/Controls/Logic/FloatInputControlLogic.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using Adobe.Substance;
 using Adobe.Substance.Input;
@@ -8,9 +6,9 @@ namespace SOS.SubstanceExtensions.Examples
 {
     public class FloatInputControlLogic : InputControlLogic<float>
     {
-        public override void Initialize(SubstanceNativeGraph nativeGraph, SubstanceGraphSO substance, ISubstanceInput input)
+        public override void Initialize(SubstanceNativeGraph nativeGraph, ISubstanceInput input)
         {
-            base.Initialize(nativeGraph, substance, input);
+            base.Initialize(nativeGraph, input);
 
             SubstanceInputFloat valueInput = (SubstanceInputFloat)input;