+++ Assets/Plugins/Substance 3D Extensions/Examples/Scripts/Runtime/Runtime/RuntimeRenderAdvancedExample.cs
@@ -39,10 +39,6 @@ namespace SOS.SubstanceExtensions.Examples
 
             //Render substance...
             await substance.RenderAsync(nativeGraph);
-
-            //Texture2D tex = null;
-            //tex.Resize(0, 0, tex.format, tex.mipmapCount > 0);
-            //tex.Reinitialize()
         }
 
 
@@ -58,7 +54,7 @@ namespace SOS.SubstanceExtensions.Examples
                 RuntimeParameterControl controlPrefab = controlsData.GetControlPrefab(inputs[i]);
                 RuntimeParameterControl control = Instantiate<RuntimeParameterControl>(controlPrefab, controlsParent);
 
-                control.Initialize(nativeGraph, substance, inputs[i]);
+                control.Initialize(nativeGraph, inputs[i]);
                 control.onValueChanged.AddListener(OnControlValueChanged);
 
                 controls.Add(control);
@@ -99,11 +95,5 @@ namespace SOS.SubstanceExtensions.Examples
 
             SetupInputControls();
         }
-
-
-        /*private void Update()
-        {
-            Debug.Log(substance.OutputRemaped);
-        }*/
     }
 }